Rename to processor handlers to be closer to composition use
